Move linear algebra function in there own module
